Decompose wolfram, add `__init__.py`
