+++ chatgpt/choose_pod.py
@@ -1,4 +1,4 @@
-from wolfram.api import Pod
+from wolfram import Pod
 
 from .config import SEED, client
 from .helper import read_prompt
