+++ wolfram/api.py
@@ -1,26 +1,11 @@
-from typing import Any
 from urllib.parse import quote_plus
-
-from pydantic import BaseModel
 
 from helper.aiohttp_client import get_client
 
 from .config import credentials
+from .types import Pod
 
 base_query = f'http://api.wolframalpha.com/v2/query?appid={credentials.APP_ID}'
-
-
-class Subpod(BaseModel):
-    title: str
-    plaintext: str
-    img: dict[str, Any]  # TODO: add typing
-
-
-class Pod(BaseModel):
-    subpods: list[Subpod]
-    id: str
-    primary: bool = False
-    title: str
 
 
 # returns list of pods
@@ -43,17 +28,3 @@ async def get_pods(query: str) -> list[Pod] | None:
 
     print(result)
     return [Pod.model_validate(pod) for pod in result['pods']]
-
-
-def extract_usefull_subpods(pod: Pod) -> list[Subpod]:
-    subpods = pod.subpods
-
-    # remove step-by-step if not available
-    subpods = [subpod for subpod in subpods if subpod.plaintext != '(step-by-step solution unavailable)']
-
-    # firstly try to extract subpods with titles
-    if with_titles := [subpod for subpod in subpods if subpod.title]:
-        return with_titles
-
-    # if no one has a title then return all of them
-    return subpods
